crawler-python/requirements.txt
kafka-python-ng
feedparser==6.0.11
trafilatura>=1.8.0
beautifulsoup4==4.12.2
lxml>=6.0.0
requests==2.31.0